+++ examples/handlebars-ad-renderer/src/tests/index.ts
@@ -8,7 +8,7 @@ import { MyHandlebarsAdRenderer } from "../MyPluginImpl";
 import { badChars, escapeChar } from "./utils";
 
 // Creative stub
-const creative: core.ResponseData<core.DisplayAd> = {
+const creative: core.DataResponse<core.DisplayAd> = {
   status: "ok",
   data: {
     type: "DISPLAY_AD",    
@@ -118,8 +118,7 @@ const templateProperties: core.AdLayoutVersionResponse = {
     template: "mics://data_file/tenants/1126/ads_templates/250.276.template",
     ad_layout_id: "250",
     status: "DRAFT"
-  },
-  count: 1
+  }
 };
 
 // Recommendation stub
@@ -298,7 +297,7 @@ describe("Test Example Handlebar Ad Renderer", function() {
     const templateContent: string = `Hello World!`;
     const rpMockup = buildRpMockup(templateContent);
 
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -322,18 +321,20 @@ describe("Test Example Handlebar Ad Renderer", function() {
               .end((err, res) => {
                 expect(res.status).to.eq(200);
 
-                expect(rpMockup
-                  .withArgs(
+                expect(
+                  rpMockup.withArgs(
                     sinon.match.has(
                       "uri",
                       sinon.match(function(value: string) {
                         return (
-                          value.match(/\/v1\/recommenders\/(.){1,10}\/recommendations/) !==
-                          null
+                          value.match(
+                            /\/v1\/recommenders\/(.){1,10}\/recommendations/
+                          ) !== null
                         );
                       })
                     )
-                  ).args[0][0].body.input_data.user_agent_id).to.be.eq(adRequest.user_agent_id);
+                  ).args[0][0].body.input_data.user_agent_id
+                ).to.be.eq(adRequest.user_agent_id);
 
                 done();
               });
@@ -347,7 +348,7 @@ describe("Test Example Handlebar Ad Renderer", function() {
     const rpMockup = buildRpMockup(templateContent);
 
     // All the magic is here
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -394,7 +395,7 @@ describe("Test Example Handlebar Ad Renderer", function() {
     const rpMockup = buildRpMockup(templateContent);
 
     // All the magic is here
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -449,7 +450,7 @@ describe("Test Example Handlebar Ad Renderer", function() {
     const rpMockup = buildRpMockup(templateContent);
 
     // All the magic is here
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -491,7 +492,7 @@ describe("Test Example Handlebar Ad Renderer", function() {
     const rpMockup = buildRpMockup(templateContent);
 
     // All the magic is here
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -532,7 +533,7 @@ describe("Test Example Handlebar Ad Renderer", function() {
     const rpMockup = buildRpMockup(templateContent);
 
     // All the magic is here
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -575,7 +576,7 @@ describe("Test Example Handlebar Ad Renderer", function() {
     const rpMockup = buildRpMockup(templateContent);
 
     // All the magic is here
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -617,7 +618,9 @@ describe("Test Example Handlebar Ad Renderer", function() {
       });
   });
 
-  it("Check that the plugin doesn't fail without any recommenderId provided", function(done) {
+  it("Check that the plugin doesn't fail without any recommenderId provided", function(
+    done
+  ) {
     // All the magic is here
 
     // Template File stub
@@ -702,20 +705,20 @@ describe("Test Example Handlebar Ad Renderer", function() {
     };
 
     rpMockup
-    .withArgs(
-      sinon.match.has(
-        "uri",
-        sinon.match(function(value: string) {
-          return (
-            value.match(/\/v1\/creatives\/(.){1,10}\/renderer_properties/) !==
-            null
-          );
-        })
+      .withArgs(
+        sinon.match.has(
+          "uri",
+          sinon.match(function(value: string) {
+            return (
+              value.match(/\/v1\/creatives\/(.){1,10}\/renderer_properties/) !==
+              null
+            );
+          })
+        )
       )
-    )
-    .returns(creativePropertiesResponse);
+      .returns(creativePropertiesResponse);
 
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // Plugin init
@@ -753,10 +756,10 @@ describe("Test Example Handlebar Ad Renderer", function() {
 
     const rpMockup = buildRpMockup(templateContent);
 
-    const plugin = new MyHandlebarsAdRenderer();
+    const plugin = new MyHandlebarsAdRenderer(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
-    const adRequest2 = Object.assign({}, adRequest); 
+    const adRequest2 = Object.assign({}, adRequest);
     adRequest2.user_agent_id = null;
 
     // Plugin init
@@ -780,23 +783,24 @@ describe("Test Example Handlebar Ad Renderer", function() {
               .end((err, res) => {
                 expect(res.status).to.eq(200);
 
-                expect(rpMockup
-                  .withArgs(
+                expect(
+                  rpMockup.withArgs(
                     sinon.match.has(
                       "uri",
                       sinon.match(function(value: string) {
                         return (
-                          value.match(/\/v1\/recommenders\/(.){1,10}\/recommendations/) !==
-                          null
+                          value.match(
+                            /\/v1\/recommenders\/(.){1,10}\/recommendations/
+                          ) !== null
                         );
                       })
                     )
-                  ).args[0][0].body.input_data.user_agent_id).to.be.eq(adRequest2.user_agent_id);
+                  ).args[0][0].body.input_data.user_agent_id
+                ).to.be.eq(adRequest2.user_agent_id);
 
                 done();
               });
           });
       });
   });
-
 });