+++ src/core/interfaces/mediarithmics/api/ad-renderer/AdRendererPluginResponseInterface.ts
@@ -1,4 +1,0 @@
-export interface AdRendererPluginResponse {
-    html: string;
-    displayContext?: any;
-  }