+++ src/core/interfaces/mediarithmics/api/email-router/EmailRouterPluginResponseInterface.ts
@@ -1,7 +1,0 @@
-export interface EmailRoutingPluginResponse {
-    result: boolean;
-}
-
-export interface CheckEmailsPluginResponse {
-    result: boolean;
-}