+++ src/core/interfaces/mediarithmics/api/recommender/RecommenderPluginResponse.ts
@@ -1,3 +1,0 @@
-import { RecommandationsWrapper } from "../common/RecommenderInterface";
-
-export interface RecommenderPluginResponse extends RecommandationsWrapper {}