+++ src/external_types/toobusy-js.d.ts
@@ -1,0 +1,1 @@
+declare module 'toobusy-js';