+++ src/index.ts
@@ -1,4 +1,4 @@
-import * as core from "./core";
+import * as core from "./mediarithmics";
 import * as extra from "./extra/handlebars";
 
 export {core, extra};