+++ src/tests/AudienceFeedConnector.ts
@@ -5,6 +5,8 @@ import * as request from "supertest";
 import * as sinon from "sinon";
 import * as mockery from "mockery";
 import * as rp from "request-promise-native";
+import {AudienceSegmentExternalFeedResource} from '../mediarithmics/api/core/audiencesegment/AudienceSegmentInterface';
+import {DataListResponse} from '../mediarithmics/api/core/common/Response';
 
 class MyFakeAudienceFeedConnector extends core.AudienceFeedConnectorBasePlugin {
   protected onExternalSegmentCreation(
@@ -46,7 +48,7 @@ const rpMockup: sinon.SinonStub = sinon.stub().returns(
 
 describe("Fetch Audience Feed Gateway API", () => {
   // All the magic is here
-  const plugin = new MyFakeAudienceFeedConnector();
+  const plugin = new MyFakeAudienceFeedConnector(false);
   const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
   it("Check that feed_id is passed correctly in fetchAudienceFeedProperties", function(
@@ -86,14 +88,14 @@ describe("Fetch Audience Feed Gateway API", () => {
 
 describe("External Audience Feed API test", function() {
   // All the magic is here
-  const plugin = new MyFakeAudienceFeedConnector();
+  const plugin = new MyFakeAudienceFeedConnector(false);
 
   it("Check that the plugin is giving good results with a simple handler", function(
     done
   ) {
     const rpMockup: sinon.SinonStub = sinon.stub();
 
-    const audienceFeed: core.ResponseData<core.AudienceFeed> = {
+    const audienceFeed: core.DataResponse<core.AudienceSegmentExternalFeedResource> = {
       status: "ok",
       data: {
         id: "74",
@@ -119,7 +121,7 @@ describe("External Audience Feed API test", function() {
     )
     .returns(audienceFeed);
 
-    const properties: core.ResponseListOfData<core.PluginProperty> = {
+    const properties: core.DataListResponse<core.PluginProperty> = {
       status: "ok",
       count: 1,
       data: [
@@ -225,26 +227,29 @@ describe("External Audience Feed API test", function() {
         expect(res.status).to.equal(200);
 
         expect(JSON.parse(res.text).status).to.be.eq("ok");
+
+        request(runner.plugin.app)
+        .post("/v1/external_segment_connection")
+        .send(externalSegmentConnection)
+        .end(function(err, res) {
+          expect(res.status).to.equal(200);
+  
+          expect(JSON.parse(res.text).status).to.be.eq("ok");
+
+          request(runner.plugin.app)
+          .post("/v1/user_segment_update")
+          .send(userSegmentUpdateRequest)
+          .end(function(err, res) {
+            expect(res.status).to.equal(200);
+    
+            expect(JSON.parse(res.text).status).to.be.eq("ok");
+    
+            done();
+          });
+
+        });
+
       });
 
-    request(runner.plugin.app)
-      .post("/v1/external_segment_connection")
-      .send(externalSegmentConnection)
-      .end(function(err, res) {
-        expect(res.status).to.equal(200);
-
-        expect(JSON.parse(res.text).status).to.be.eq("ok");
-      });
-
-    request(runner.plugin.app)
-      .post("/v1/user_segment_update")
-      .send(userSegmentUpdateRequest)
-      .end(function(err, res) {
-        expect(res.status).to.equal(200);
-
-        expect(JSON.parse(res.text).status).to.be.eq("ok");
-
-        done();
-      });
   });
 });